fixing the learn spell twice problem, made itemuse actions threadsafe, because now they share some state

+++ src/main/java/com/comandante/creeper/Items/use/LightningSpellBookUseAction.java
@@ -4,18 +4,10 @@ import com.comandante.creeper.Items.Item;
 import com.comandante.creeper.Items.ItemType;
 import com.comandante.creeper.Items.ItemUseAction;
 import com.comandante.creeper.managers.GameManager;
-import com.comandante.creeper.npc.Npc;
-import com.comandante.creeper.npc.NpcStatsChangeBuilder;
 import com.comandante.creeper.player.Player;
-import com.comandante.creeper.server.Color;
 import com.comandante.creeper.spells.Effect;
 import com.comandante.creeper.spells.LightningSpell;
-import com.comandante.creeper.stat.StatsBuilder;
-import com.comandante.creeper.world.Room;
 
-import java.text.NumberFormat;
-import java.util.Arrays;
-import java.util.Locale;
 import java.util.Set;
 
 public class LightningSpellBookUseAction implements ItemUseAction {
@@ -26,6 +18,8 @@ public class LightningSpellBookUseAction implements ItemUseAction {
         this.itemType = itemType;
     }
 
+    private Boolean dontDelete = Boolean.FALSE;
+
     @Override
     public Integer getItemTypeId() {
         return itemType.getItemTypeCode();
@@ -33,14 +27,21 @@ public class LightningSpellBookUseAction implements ItemUseAction {
 
     @Override
     public void executeAction(GameManager gameManager, Player player, Item item) {
+        if (player.getLearnedSpells().contains(LightningSpell.NAME)) {
+            gameManager.getChannelUtils().write(player.getPlayerId(), "You already know how to use " + LightningSpell.NAME);
+            dontDelete = true;
+            return;
+        }
         gameManager.writeToPlayerCurrentRoom(player.getPlayerId(), player.getPlayerName() + " reads a leatherbound aging spell book and gains knowledge about lightning spells.");
         player.addLearnedSpellByName(LightningSpell.NAME);
     }
 
     @Override
     public void postExecuteAction(GameManager gameManager, Player player, Item item) {
-        player.removeInventoryId(item.getItemId());
-        gameManager.getEntityManager().removeItem(item);
+        if (!dontDelete) {
+            player.removeInventoryId(item.getItemId());
+            gameManager.getEntityManager().removeItem(item);
+        }
     }
 
     @Override
